LayoutTests/platform/chromium/fast/events/touch/gesture/touch-gesture-scroll-page-expected.txt

This tests scroll gesture event scrolling on a whole page.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


first gesture
wheel event 0+> [object HTMLDivElement]
wheel event 1+> [object HTMLDivElement]
PASS document.body.scrollTop is 70
PASS document.body.scrollLeft is 0
PASS wheelEventsOccurred is 2
second gesture
wheel event 0+> [object HTMLDivElement]
wheel event 1+> [object HTMLDivElement]
PASS document.body.scrollTop is 130
PASS document.body.scrollLeft is 0
PASS wheelEventsOccurred is 2
scroll event 0+> [object HTMLDocument]
PASS scrollEventsOccurred is 1
PASS successfullyParsed is true

TEST COMPLETE
